refactor: move inner class
